DynamicalNetworks.jl

DynamicalNetworks.Lorenz — Type

struct Lorenz{T1, T2, T3, T4} <: NodeDynamics

Lorenz dynamics that evolves by

\[\dot{x}_{1} = \sigma (x_2 - x_1) \ \dot{x}_{2} = x_1 (\rho - x_3) - x_2 \ \dot{x}_{3} = x_1 x_2 - \beta x_3\]

where $\sigma, \beta, \rho$ are system parameters.

Fields

σ::Any

β::Any

ρ::Any

x::Any

DynamicalNetworks.Network — Type

struct Network{T1, T2, T3}

A network consisting of dynamical systems. The dynamics of the network evolves by,

\[ \dot{x}_i = f_i(x_i) + \sum_{j = 1}^{n} \epsilon_{ij} P x_j \quad i = 1, \ldots, n\]

where $x_i$ is the state vector of node $i$, $f_i'' is the individual node dynamics,$\epsilon{ỉj} \geq 0$is the coupling strength between the nodes$i$and$j$.$P = diag(p1, \ldots, p_d)`` determines the state variables by which the nodes are coupled.

Fields

nodes::Any

E::Any

P::Any

DynamicalNetworks.Simulation — Type

struct Simulation

A simlation object that containining simulation info such as the path of the simulation data file and simulation status.

Fields

filepath::String

retcode::Symbol

DynamicalNetworks.readsim — Method

readsim(sim)

Reads the simulation data and returns a tuple of simulation time and simulation data.

DynamicalNetworks.signalflow — Method

signalflow(net)

Plots signal flow graph.

DynamicalNetworks.simulate — Function

simulate(net)
simulate(net, ti)
simulate(net, ti, dt)
simulate(net, ti, dt, tf)
simulate(net, ti, dt, tf, solargs; solkwargs, path, filename)

Simulates net from ti to tf with a step size ofdt`.

DynamicalNetworks.getprob — Method

getprob(net, tspan)

Returns an ODE problem for a time span of tspan.

DynamicalNetworks.Lorenz

DynamicalNetworks.Network

DynamicalNetworks.Simulation

DynamicalNetworks.getprob

DynamicalNetworks.readsim

DynamicalNetworks.signalflow

DynamicalNetworks.simulate